refactor: StudentSelectModel changed to module
index.html
- remove script tag for StudentSelectModel.js

StudentSelectModel.js
- removed the old script file

modules/StudentSelectPageModel
- new ES6 module file

sw.js
- update cache to cache new module file

app.js
- import StudentSelectPageModel
