+++ src/html/index.html
@@ -86,7 +86,6 @@
     </body>
     
     <script type="application/javascript" charset="utf-8" src="scripts/idb.js"></script>
-    <script type="application/javascript" charset="utf-8" src="scripts/StudentSelectModel.js"></script>
     <script type="application/javascript" charset="utf-8" src="scripts/ServiceWorkerHelper.js"></script>
     <script type="module">
       import app from "/scripts/app.js";
